requirements.txt
requests~=2.32.3
tqdm~=4.67.1
python-dotenv~=1.0.1
pytest~=8.3.4
fusepy~=3.0.1